aegis_ai/requirements.txt
Django==5.1.*
djangorestframework
scikit-learn
pandas
numpy
pdfminer.six
pytesseract
pyzbar
Pillow
python-dotenv